+++ packages/components/scripts/generateDSRComponents.ts
@@ -261,15 +261,22 @@ import { get${componentName}Css } from '${stylesBundleImportPath}';
       }
 
       if (!newFileContent.includes('export const InputBase:')) {
-        newFileContent = newFileContent
-          .replace(
+        // radio-group-option uses a label component without allowing slots
+        if (tagName === 'p-radio-group-option') {
+          newFileContent = newFileContent.replace(
+            /(<Label(?!Props))([\s\S]*?\/>)/,
+            '$1 hasLabel={this.props.label} hasDescription={false}$2'
+          );
+        } else {
+          newFileContent = newFileContent.replace(
             /(<Label(?!Props))([\s\S]*?\/>)/,
             "$1 hasLabel={this.props.label || namedSlotChildren.filter(({ props: { slot } }) => slot === 'label').length > 0} hasDescription={this.props.description || namedSlotChildren.filter(({ props: { slot } }) => slot === 'description').length > 0}$2"
-          )
-          .replace(
-            /(<StateMessage(?!Props))([\s\S]*?\/>)/,
-            "$1 hasMessage={(this.props.message || namedSlotChildren.filter(({ props: { slot } }) => slot === 'message').length > 0) && ['success', 'error'].includes(this.props.state)}$2"
           );
+        }
+        newFileContent = newFileContent.replace(
+          /(<StateMessage(?!Props))([\s\S]*?\/>)/,
+          "$1 hasMessage={(this.props.message || namedSlotChildren.filter(({ props: { slot } }) => slot === 'message').length > 0) && ['success', 'error'].includes(this.props.state)}$2"
+        );
       } else {
         newFileContent = newFileContent
           .replace(
@@ -601,6 +608,15 @@ import { get${componentName}Css } from '${stylesBundleImportPath}';
           .replace(/<span className="sr-only"[^<]*<\/span>/, '')
           // .replace(/(SelectDropdownDirectionInternal)/, 'type $1')
           .replace(/private searchTimeout: any\.Timeout \| number = null;/, '')
+          // TODO replace ElementInternals lifecycle callbacks (formAssociatedCallback, formDisabledCallback, formResetCallback, formStateRestoreCallback) completely
+          .replace(/@AttachInternals\(\)/, '')
+          .replace(/this\.props\.value = this\.props\.defaultValue;/, '')
+          .replace(/this\.props\.disabled = disabled;/, '')
+          .replace(/this\.props\.value = state;/, '')
+          .replace(/formDisabledCallback\(disabled: boolean\)/, 'formDisabledCallback()')
+          .replace(/formStateRestoreCallback\(state: string\)/, 'formStateRestoreCallback()');
+      } else if (tagName === 'p-radio-group') {
+        newFileContent = newFileContent
           // TODO replace ElementInternals lifecycle callbacks (formAssociatedCallback, formDisabledCallback, formResetCallback, formStateRestoreCallback) completely
           .replace(/@AttachInternals\(\)/, '')
           .replace(/this\.props\.value = this\.props\.defaultValue;/, '')