+++ packages/components/CHANGELOG.md
@@ -14,6 +14,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ## [Unreleased]
 
+### Added
+
+- `Pin Code`, `Segmented Control`:
+  - `update` event is deprecated. Use `change` event instead.
+  - `blur` event ([#4050](https://github.com/porsche-design-system/porsche-design-system/pull/4050))
+- `Select`, `Multi Select`:
+  - `update` event is deprecated. Use `change` event instead.
+  - `blur` event ([#4046](https://github.com/porsche-design-system/porsche-design-system/pull/4046))
+- `Radio Group`: ([#4042](https://github.com/porsche-design-system/porsche-design-system/pull/4042))
+- `Radio Group`: `blur` event ([#4057](https://github.com/porsche-design-system/porsche-design-system/pull/4057))
+
+### Changed
+
+- `Radio Button Wrapper`: component is deprecated. Use `Radio Group` instead.
+  ([#4042](https://github.com/porsche-design-system/porsche-design-system/pull/4042))
+
+
+### Fixed
+
+- `Select Wrapper`: disabled options in native dropdown not displayed correctly
+  ([#4046](https://github.com/porsche-design-system/porsche-design-system/pull/4046))
+- `Tailwind`: change `--color-base` to `--color-canvas` to avoid conflicts with `text-base`
+
 ## [3.30.0-rc.0] - 2025-09-10
 
 ### Added